refactor: Move DP calibration onto the noise sampler
forDifferentialPrivacy(epsilon, delta, sensitivity) now lives on
DeterministicTruncatedLaplaceNoiseSampler, which owns DP noise generation.
TruncatedLaplaceNoiseDistribution is left a pure (scale, bound) primitive with no
differential-privacy concepts. The noiser calibrates via the sampler factory.

Move the calibration unit tests to the sampler test; the distribution test now
covers only the primitive.
